requirements.txt
py-cord
python-dotenv
black
flake8
